Cleans up some packets
