+++ Projects/Server/Network/Packets/Old Packets/PlayerPackets.cs
@@ -49,14 +49,6 @@ namespace Server.Network
     }
   }
 
-  public sealed class ChangeCombatant : Packet
-  {
-    public ChangeCombatant(Mobile combatant) : base(0xAA, 5)
-    {
-      Stream.Write(combatant?.Serial ?? Serial.Zero);
-    }
-  }
-
   public sealed class ChangeUpdateRange : Packet
   {
     private static readonly ChangeUpdateRange[] m_Cache = new ChangeUpdateRange[0x100];
@@ -94,21 +86,6 @@ namespace Server.Network
     public static Packet Instantiate(bool dead) => dead ? Dead : Alive;
   }
 
-  public sealed class SpeedControl : Packet
-  {
-    public static readonly Packet WalkSpeed = SetStatic(new SpeedControl(2));
-    public static readonly Packet MountSpeed = SetStatic(new SpeedControl(1));
-    public static readonly Packet Disable = SetStatic(new SpeedControl(0));
-
-    public SpeedControl(int speedControl) : base(0xBF)
-    {
-      EnsureCapacity(3);
-
-      Stream.Write((short)0x26);
-      Stream.Write((byte)speedControl);
-    }
-  }
-
   public sealed class ToggleSpecialAbility : Packet
   {
     public ToggleSpecialAbility(int abilityID, bool active) : base(0xBF)
@@ -119,48 +96,6 @@ namespace Server.Network
 
       Stream.Write((short)abilityID);
       Stream.Write(active);
-    }
-  }
-
-  public sealed class GlobalLightLevel : Packet
-  {
-    private static readonly GlobalLightLevel[] m_Cache = new GlobalLightLevel[0x100];
-
-    public GlobalLightLevel(int level) : base(0x4F, 2)
-    {
-      Stream.Write((sbyte)level);
-    }
-
-    public static GlobalLightLevel Instantiate(int level)
-    {
-      var lvl = (byte)level;
-      var p = m_Cache[lvl];
-
-      if (p == null)
-      {
-        m_Cache[lvl] = p = new GlobalLightLevel(level);
-        p.SetStatic();
-      }
-
-      return p;
-    }
-  }
-
-  public sealed class PersonalLightLevel : Packet
-  {
-    public PersonalLightLevel(Mobile m, int level) : base(0x4E, 6)
-    {
-      Stream.Write(m.Serial);
-      Stream.Write((sbyte)level);
-    }
-  }
-
-  public sealed class PersonalLightLevelZero : Packet
-  {
-    public PersonalLightLevelZero(Mobile m) : base(0x4E, 6)
-    {
-      Stream.Write(m.Serial);
-      Stream.Write((sbyte)0);
     }
   }
 
@@ -204,61 +139,6 @@ namespace Server.Network
       Stream.Write((byte)type);
       Stream.Write((byte)density);
       Stream.Write((byte)temp);
-    }
-  }
-
-  /// <summary>
-  ///   Causes the client to walk in a given direction. It does not send a movement request.
-  /// </summary>
-  public sealed class MovePlayer : Packet
-  {
-    public MovePlayer(Direction d) : base(0x97, 2)
-    {
-      Stream.Write((byte)d);
-
-      // @4C63B0
-    }
-  }
-
-  public sealed class SetWarMode : Packet
-  {
-    public static readonly Packet InWarMode = SetStatic(new SetWarMode(true));
-    public static readonly Packet InPeaceMode = SetStatic(new SetWarMode(false));
-
-    public SetWarMode(bool mode) : base(0x72, 5)
-    {
-      Stream.Write(mode);
-      Stream.Write((byte)0x00);
-      Stream.Write((byte)0x32);
-      Stream.Write((byte)0x00);
-      // m_Stream.Fill();
-    }
-
-    public static Packet Instantiate(bool mode) => mode ? InWarMode : InPeaceMode;
-  }
-
-  public sealed class Swing : Packet
-  {
-    public Swing(int flag, Mobile attacker, Mobile defender) : base(0x2F, 10)
-    {
-      Stream.Write((byte)flag);
-      Stream.Write(attacker.Serial);
-      Stream.Write(defender.Serial);
-    }
-  }
-
-  public sealed class NullFastwalkStack : Packet
-  {
-    public NullFastwalkStack() : base(0xBF)
-    {
-      EnsureCapacity(256);
-      Stream.Write((short)0x1);
-      Stream.Write(0x0);
-      Stream.Write(0x0);
-      Stream.Write(0x0);
-      Stream.Write(0x0);
-      Stream.Write(0x0);
-      Stream.Write(0x0);
     }
   }
 
@@ -370,8 +250,7 @@ namespace Server.Network
 
   public sealed class SeasonChange : Packet
   {
-    private static readonly SeasonChange[][] m_Cache = new SeasonChange[][]
-    {
+    private static readonly SeasonChange[][] m_Cache = {
       new SeasonChange[2],
       new SeasonChange[2],
       new SeasonChange[2],
@@ -527,44 +406,6 @@ namespace Server.Network
       if (p == null)
       {
         m_Cache[ping] = p = new PingAck(ping);
-        p.SetStatic();
-      }
-
-      return p;
-    }
-  }
-
-  public sealed class MovementRej : Packet
-  {
-    public MovementRej(int seq, Mobile m) : base(0x21, 8)
-    {
-      Stream.Write((byte)seq);
-      Stream.Write((short)m.X);
-      Stream.Write((short)m.Y);
-      Stream.Write((byte)m.Direction);
-      Stream.Write((sbyte)m.Z);
-    }
-  }
-
-  public sealed class MovementAck : Packet
-  {
-    private static readonly MovementAck[] m_Cache = new MovementAck[8 * 256];
-
-    private MovementAck(int seq, int noto) : base(0x22, 3)
-    {
-      Stream.Write((byte)seq);
-      Stream.Write((byte)noto);
-    }
-
-    public static MovementAck Instantiate(int seq, Mobile m)
-    {
-      var noto = Notoriety.Compute(m, m);
-
-      var p = m_Cache[noto * seq];
-
-      if (p == null)
-      {
-        m_Cache[noto * seq] = p = new MovementAck(seq, noto);
         p.SetStatic();
       }
